Add a 'jump to nearest' button
This adds a button to the main view to jump to the nearest tube
station. The current location is queried from the webOS location
service. A separate Javascript file is added which contains a
hard-coded list of coordinates for each station along with a pixel
position inside the image. The scroller is scrolled to display the
correct station. A banner div is added to the main view to report how
far the station is away.